Refactoring/cleaning of the model-chain management/API and it's interaction with the underlying IO-messaging code.
More transparent handling of the underlying IO-code but without exposing the ASIO-specific mechanisms (for cases if/when different IO-related code is to be deployed, such as custom BSD-centric 'kqueue' implementation of single-threaded messaging which may/will guarantee that the act of IO-cancelling will not deliver any callbacks afterwards).

Somewhat cleaner code (deprecation of 'write_more', 'being_taken_over', 'final_in_chain' variables etc. etc.)

Template-refactoring of io_wrapper for the preparation for more generalised IO-messaging (in future, capable of handling different types of async_drivers such as textual readline variants for the purpose of web-like server functionality and so on).

Beginning to work on the task management such as task-completion -cancellation (i.e. without assert(0)/crashing-on-purpose when a given convergence completes).

More robust handling of the 'end_of_process' messaging sequence (e.g. skipping stray messages in the pipeline, etc.)

Deprecating some of the 'to be deprecated' code (e.g. swapping censoc::ptr for the ::boost::scoped_ptr and ::std::auto_ptr usage).