Task id/name reworking w.r.t. html display -- somewhat friendlier notation (displaying the model name such as Logit, MixL, etc. and culling irrelavant numeric 'noise' :-) givan that the first digits (task id) are the only uniquely identifying numbers in a given server's deployment).
Refactoring of the "short-description" field to, now, include the multiline log of the sibmission parameters when adding a new task... rather a quick hack and  will probably need to be reworked (e.g. at the very least the title from 'short description' to something else... as it's not really that short now...

Minor CSS re-styling (buttons, checkboxes, et. al.)