Bug-fix (server crash in release mode) -- moving static numpunct specialization (coma-separated thousands) facet to a separate translation unit. This is done mainly to allow the overall building process to continue using the "-fno-rtti" option (in release mode) whilst at the same time generating compatible rtti typeinfos for objects (i.e. the facet in question) being passed/managed by a separately compiled (with rtti) libstdc++ binary object(s).
Subsequent refactoring of the interface for the lexicast locale/imbue to make it more stl-like. Also, to prevent multiple definition for the same functions and as a matter of efficcient-related pedanticity -- making some of the lexicast.h functions declared as static.

The reasons for one to insist on 'no rtti' globally is due to the following: in the spirit of c++ language design, if one does not want to use 'dynamic_cast' et. al. then one should not pay the price for such a feature(s). One the one hand such a price may be quite smallish -- w.r.t. additional memory usage and possible reduction of CPUs cache-compactness. On the other hand, there could be noticeable performance degradation as hash-defs in some lib-headers may, indeed, detect rtti presence and chose to take a somewhat slower-path-of-execution for the code flow... this may include the headers from some c++ libs which do allow linking from non-rtti enabled programs to the rtti-enabled libs...