Another "snapshot of a commit" -- once again, not sure when I'll be able to look at this project again so "syncing" any of the local changes (could break the build/compile stages, etc. -- use a few revisions back if need be).
Minor refactoring of the arch. in http subdir (allows for http-subset functionality to be built as a standalone(ish) server in future).

Commit of Luke's contirbutions to the "install and use" howto of "autoinput" (a small project in the leon_wip). Ideally I was going to convert the file to LaTeX but never got the time to do this, so committing "as is".